Cretae util function for date format
